Refactor Codex turn verification evidence helpers
